requirements.txt
openai-agents
unitycatalog-openai[databricks]==0.2.0
databricks-agents
mlflow==2.21.0
uv